Work around resource group scrubbing bug.
