+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/tests/latest/test_vm_commands.py
@@ -3147,6 +3147,81 @@ class VMGalleryImage(ScenarioTest):
 
 # endregion
 
+# region ppg tests
+
+class ProximityPlacementGroupScenarioTest(ScenarioTest):
+
+    @ResourceGroupPreparer(name_prefix="cli_test_ppg_cmds_")
+    def test_proximity_placement_group(self, resource_group, resource_group_location):
+        self.kwargs.update({
+            'ppg1': 'my_ppg_1',
+            'ppg2': 'my_ppg_2',
+            'loc': resource_group_location
+        })
+
+        self.cmd('ppg create -n {ppg1} -t standard -g {rg}', checks=[
+            self.check('name', '{ppg1}'),
+            self.check('location', '{loc}'),
+            self.check('proximityPlacementGroupType', 'Standard')
+        ])
+
+        self.cmd('ppg create -n {ppg2} -t ultra -g {rg}', checks=[
+            self.check('name', '{ppg2}'),
+            self.check('location', '{loc}'),
+            self.check('proximityPlacementGroupType', 'Ultra')
+        ])
+
+    @ResourceGroupPreparer(name_prefix='cli_test_ppg_resources_')
+    def test_ppg_with_related_resources(self, resource_group):
+
+        self.kwargs.update({
+            'ppg': 'my_ppg_1',
+            'vm': 'vm1',
+            'vmss': 'vmss1',
+            'avset': 'avset1',
+            'ssh_key': TEST_SSH_KEY_PUB
+        })
+
+        self.kwargs['ppg_id'] = self.cmd('ppg create -n {ppg} -t standard -g {rg}').get_output_in_json()['id']
+
+        self.kwargs['vm_id'] = self.cmd('vm create -g {rg} -n {vm} --image UbuntuLTS --ssh-key-value \'{ssh_key}\' --ppg {ppg}').get_output_in_json()['id']
+
+        self.cmd('vmss create -g {rg} -n {vmss} --image UbuntuLTS --ssh-key-value \'{ssh_key}\' --ppg {ppg_id}')
+        self.kwargs['vmss_id'] = self.cmd('vmss show -g {rg} -n {vmss}').get_output_in_json()['id']
+
+        self.kwargs['avset_id'] = self.cmd('vm availability-set create -g {rg} -n {avset} --ppg {ppg}').get_output_in_json()['id']
+
+        ppg_resource = self.cmd('ppg show -n {ppg} -g {rg}').get_output_in_json()
+
+        # check that the compute resources are created with PPG
+
+        self._assert_ids_equal(ppg_resource['availabilitySets'][0]['id'], self.kwargs['avset_id'], rg_prefix='cli_test_ppg_resources_')
+        self._assert_ids_equal(ppg_resource['virtualMachines'][0]['id'], self.kwargs['vm_id'], rg_prefix='cli_test_ppg_resources_')
+        self._assert_ids_equal(ppg_resource['virtualMachineScaleSets'][0]['id'], self.kwargs['vmss_id'], 'cli_test_ppg_resources_')
+
+
+    # it would be simpler to do the following:
+    # self.assertEqual(ppg_resource['availabilitySets'][0]['id'].lower(), self.kwargs['avset_id'].lower())
+    # self.assertEqual(ppg_resource['virtualMachines'][0]['id'].lower(), self.kwargs['vm_id'].lower())
+    # self.assertEqual(ppg_resource['virtualMachineScaleSets'][0]['id'].lower(), self.kwargs['vmss_id'].lower())
+    #
+    # however, the CLI does not replace resource group values in payloads in the recordings.
+    def _assert_ids_equal(self, id_1, id_2, rg_prefix=None):
+        from msrestazure.tools import parse_resource_id
+
+        id_1, id_2, rg_prefix = id_1.lower(), id_2.lower(), rg_prefix.lower() if rg_prefix else rg_prefix
+
+        parsed_1, parsed_2 = parse_resource_id(id_1), parse_resource_id(id_2)
+
+        self.assertEqual(len(parsed_1.keys()), len(parsed_2.keys()))
+
+        for k1, k2 in zip(sorted(parsed_1.keys()), sorted(parsed_2.keys())):
+            if rg_prefix is not None and k1 == 'resource_group':
+                self.assertTrue(parsed_1[k1].startswith(rg_prefix))
+                self.assertTrue(parsed_2[k2].startswith(rg_prefix))
+            else:
+                self.assertEqual(parsed_1[k1], parsed_2[k2])
+
 
 if __name__ == '__main__':
     unittest.main()
